# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Historial Clinico Backend",
        "description": "Aplicación backend del sistema web Historial Clnico",
        "contact": {
            "email": "dav.almeyda@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/historial-clinico-backend/public/api/admin/getEspecialidades": {
            "post": {
                "tags": [
                    "Administrador"
                ],
                "summary": "Obtener las Especialidades",
                "operationId": "getEspecialidades",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "codGrupoCia": {
                                        "type": "string"
                                    },
                                    "codLocal": {
                                        "type": "string"
                                    },
                                    "codMedico": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "codGrupoCia": "001",
                                    "codLocal": "001",
                                    "codMedico": "0000026144"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Datos Encontrados"
                    }
                }
            }
        },
        "/historial-clinico-backend/public/api/admin/getListaAtenciones": {
            "post": {
                "tags": [
                    "Administrador"
                ],
                "summary": "Obtener lista de Atenciones",
                "operationId": "getListaAtenciones",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "codGrupoCia": {
                                        "type": "string"
                                    },
                                    "codMedico": {
                                        "type": "string"
                                    },
                                    "fechaInicio": {
                                        "type": "string"
                                    },
                                    "fechaFin": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "codGrupoCia": "001",
                                    "codMedico": "0000026144",
                                    "fechaInicio": "01/11/2021",
                                    "fechaFin": "01/12/2021"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Datos Encontrados"
                    }
                }
            }
        },
        "/historial-clinico-backend/public/api/admin/getListaLiberados": {
            "post": {
                "tags": [
                    "Administrador"
                ],
                "summary": "Obtener lista de Liberados",
                "operationId": "getListaLiberados",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "codGrupoCia": {
                                        "type": "string"
                                    },
                                    "fechaInicio": {
                                        "type": "string"
                                    },
                                    "fechaFin": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "codGrupoCia": "001",
                                    "fechaInicio": "01/11/2021",
                                    "fechaFin": "01/12/2021"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Datos Encontrados"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}